Dist

Description

LAoP is a library for algebraic (inductive) construction and manipulation of matrices in Haskell. See my Msc Thesis for the motivation behind the library, the underlying theory, and implementation details.

This module exports a type synonym Dist that represents a stochastic distribution matrix and several distribution construction functions.

Documentation

If the sum of the rows of a column matrix is equal to 1 then this stochastic matrix can be seen as a probability distribution.

This module is still experimental but it's already possible to model probabilistic programming problems with it. Import Nat or Type to access LAoP matrix combinators and then all you have to do is to define your sample space, either by creating a new data type or by abstracting it out via Natural.

Write manipulation functions and promote them to matrices via fromF or fromF' and you're good to go!

Dist and Prob type synonyms

newtype Dist a

Type synonym for column vector matrices. This represents a probability distribution.

Constructors

D (Matrix Prob () a)

type Prob = Double

Type synonym for probability value

Functor instance equivalent functions

fmapD :: (Bounded a, Bounded b, Enum a, Enum b, Eq b, KnownNat (Count (Normalize a)), KnownNat (Count (Normalize b)), FromLists Prob (Normalize b) (Normalize a)) => (a -> b) -> Dist a -> Dist b

Functor instance

Applicative equivalent functions

unitD :: Dist ()

Applicative/Monoidal instance unit function

multD :: (KnownNat (Count (Normalize a)), KnownNat (Count (Normalize b)), KnownNat (Count (Normalize (a, b))), FromLists Prob (Normalize (a, b)) (Normalize a), FromLists Prob (Normalize (a, b)) (Normalize b), Normalize (a, b) ~ Normalize (Normalize a, Normalize b)) => Dist a -> Dist b -> Dist (a, b)

Applicative/Monoidal instance mult function

Selective equivalent functions

selectD :: (Normalize (Either a b) ~ Either (Normalize a) (Normalize b), FromLists Prob (Normalize b) (Normalize b), KnownNat (Count (Normalize b))) => Dist (Either a b) -> Matrix Prob a b -> Dist b

Selective instance function

Monad equivalent functions

returnD :: forall a. (Enum a, FromLists Prob () (Normalize a), KnownNat (Count a)) => a -> Dist a

Monad instance return function

bindD :: Dist a -> Matrix Prob a b -> Dist b

Monad instance '(>>=)' function

Distribution construction functions

choose :: FromLists Prob () (Normalize a) => Prob -> Dist a

Constructs a Bernoulli distribution

shape :: FromLists Prob () (Normalize a) => (Prob -> Prob) -> [a] -> Dist a

Creates a distribution given a shape function

linear :: FromLists Prob () (Normalize a) => [a] -> Dist a

Constructs a Linear distribution

uniform :: FromLists Prob () (Normalize a) => [a] -> Dist a

Constructs an Uniform distribution

negExp :: FromLists Prob () (Normalize a) => [a] -> Dist a

Constructs an Negative Exponential distribution

normal :: FromLists Prob () (Normalize a) => [a] -> Dist a

Constructs an Normal distribution

Converto to list of pairs

toValues :: forall a. (Enum a, KnownNat (Count a), FromLists Prob () (Normalize a)) => Dist a -> [(a, Prob)]

Transforms a Dist into a list of pairs.

Pretty print distribution

prettyDist :: forall a. (Show a, Enum a, KnownNat (Count a), FromLists Prob () (Normalize a)) => Dist a -> String

Pretty a distribution

prettyPrintDist :: forall a. (Show a, Enum a, KnownNat (Count a), FromLists Prob () (Normalize a)) => Dist a -> IO ()

Pretty Print a distribution